Moved EvalMonitorHook in tensorflow sub-package
